lesson3/requirements.txt
fastapi==0.89.1
uvicorn==0.22.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pymongo==4.4.1